Xyllomer 1.4.5

+++ src/config.h
@@ -17,8 +17,8 @@
  */
 
 /* these may be changed, but sizeof(type) <= sizeof(int) */
-typedef unsigned short uindex;
-# define UINDEX_MAX	USHRT_MAX
+typedef unsigned int  uindex;
+# define UINDEX_MAX	UINT_MAX
 
 typedef uindex sector;
 # define SW_UNUSED	UINDEX_MAX
